align project structure
