--- cluster/cluster.go
@@ -1,5 +1,0 @@
-package cluster
-
-type Cluster struct {
-	ID string `json:"id" yaml:"id"`
-}
