--- customer/customer.go
@@ -1,5 +1,0 @@
-package customer
-
-type Customer struct {
-	ID string `json:"id" yaml:"id"`
-}
