--- docker/daemon/daemon.go
@@ -1,5 +1,0 @@
-package daemon
-
-type Daemon struct {
-	ExtraArgs string `json:"extraArgs" yaml:"extraArgs"`
-}
