--- docker/registry/registry.go
@@ -1,5 +1,0 @@
-package registry
-
-type Registry struct {
-	Endpoint string `json:"endpoint" yaml:"endpoint"`
-}
