--- kubernetes/dns/dns.go
@@ -1,10 +1,0 @@
-package dns
-
-import (
-	"net"
-)
-
-type DNS struct {
-	// IP is the kube DNS cluster IP, e.g. 172.31.0.10.
-	IP net.IP `json:"ip" yaml:"ip"`
-}
