--- kubernetes/hyperkube/hyperkube.go
@@ -1,9 +1,0 @@
-package hyperkube
-
-import (
-	"github.com/giantswarm/clustertpr/kubernetes/hyperkube/docker"
-)
-
-type Hyperkube struct {
-	Docker docker.Docker `json:"docker" yaml:"docker"`
-}
