--- kubernetes/kubectl/kubectl.go
@@ -1,7 +1,0 @@
-package kubectl
-
-import "github.com/giantswarm/clustertpr/kubernetes/kubectl/docker"
-
-type Kubectl struct {
-	Docker docker.Docker `json:"docker" yaml:"docker"`
-}
