--- node/node.go
@@ -1,5 +1,0 @@
-package node
-
-type Node struct {
-	ID string `json:"id" yaml:"id"`
-}
