+++ spec/cluster.go
@@ -1,0 +1,5 @@
+package spec
+
+type Cluster struct {
+	ID string `json:"id" yaml:"id"`
+}
