+++ spec/customer.go
@@ -1,0 +1,5 @@
+package spec
+
+type Customer struct {
+	ID string `json:"id" yaml:"id"`
+}
