+++ spec/docker/daemon.go
@@ -1,0 +1,5 @@
+package docker
+
+type Daemon struct {
+	ExtraArgs string `json:"extraArgs" yaml:"extraArgs"`
+}
