+++ spec/docker/registry.go
@@ -1,0 +1,5 @@
+package docker
+
+type Registry struct {
+	Endpoint string `json:"endpoint" yaml:"endpoint"`
+}
