+++ spec/kubernetes/dns.go
@@ -1,0 +1,10 @@
+package kubernetes
+
+import (
+	"net"
+)
+
+type DNS struct {
+	// IP is the kube DNS cluster IP, e.g. 172.31.0.10.
+	IP net.IP `json:"ip" yaml:"ip"`
+}
