+++ spec/kubernetes/hyperkube.go
@@ -1,0 +1,7 @@
+package kubernetes
+
+import "github.com/giantswarm/clustertpr/spec/kubernetes/hyperkube"
+
+type Hyperkube struct {
+	Docker hyperkube.Docker `json:"docker" yaml:"docker"`
+}
