+++ spec/kubernetes/kubectl.go
@@ -1,0 +1,7 @@
+package kubernetes
+
+import "github.com/giantswarm/clustertpr/spec/kubernetes/kubectl"
+
+type Kubectl struct {
+	Docker kubectl.Docker `json:"docker" yaml:"docker"`
+}
