+++ spec/node.go
@@ -1,0 +1,5 @@
+package spec
+
+type Node struct {
+	ID string `json:"id" yaml:"id"`
+}
