+++ spec/vault.go
@@ -1,0 +1,6 @@
+package spec
+
+type Vault struct {
+	Address string `json:"address" yaml:"address"`
+	Token   string `json:"token" yaml:"token"`
+}
