--- vault/vault.go
@@ -1,6 +1,0 @@
-package vault
-
-type Vault struct {
-	Address string `json:"address" yaml:"address"`
-	Token   string `json:"token" yaml:"token"`
-}
